Chore: refactor to allow new pages and component reuse
